Week 07 Homework
reduced the size of the container margins to 94% so the content did not
overlap and become one column.

Played around with font sizes and margins to create a responsive layout
appropriate for 768px wide.

Hide the ad unit when creating a single-column layout.

Reverse the positioning of the div’s in the aside so that categories
appear just before the footer — the thinking behind this was that
people tend to scroll until they see links, then assume the rest is
irrelevant/ more links.

There is code in here that is commented out - I was going to make a
hamburger menu but ran out of time to finish it off and apply jquery.